+++ apps/pages/migrations/0007_page_hero_image.py
@@ -14,6 +14,6 @@ class Migration(migrations.Migration):
         migrations.AddField(
             model_name='page',
             name='hero_image',
-            field=models.ImageField(blank=True, null=True, upload_to='pages/hero/', validators=[apps.core.validators.ImageValidator(max_height=2500, max_kb=300, max_width=2500, min_height=400, min_width=400)]),
+            field=models.ImageField(blank=True, null=True, upload_to='pages/hero/', validators=[apps.core.validators.ImageValidator(max_height=2500, max_width=2500, min_height=400, min_width=400)]),
         ),
     ]